Promotion UI Complete Chess
Chess is complete with promotion ui in the project.

Gonna be making sure it works well on mobile, ui is locked and move list is set to a scroll rect, testing it, and then implimenting minimax algorithm for ai and evaluation bar. Stoked, youtube video coming up soon to go over the changes.